requirements.txt
python-telegram-bot==21.6
yfinance==0.2.40
